requirements.txt
flask
requests
python-whois
colorama
tabulate
jinja2
pyyaml
speechrecognition
pyaudio
yara-python
geoip2
python-dotenv
werkzeug
reportlab
pillow
pytesseract
pdfplumber
elasticsearch>=8.0.0
elasticsearch-dsl>=8.0.0
flask-socketio
python-socketio
eventlet
